+++ src/components/Header/Header.js
@@ -26,6 +26,13 @@ import { logoutUser } from 'actions/auth';
 import { toggleSidebar, openSidebar, closeSidebar, changeActiveSidebarItem } from '../../actions/navigation';
 
 import adminDefault from '../../images/people/chat2.png';
+import MenuIcon from '../../images/sidebar/Fill/MenuIcon';
+import FlipIcon from '../../images/sidebar/Outline/Flip';
+import CloseIcon from '../../images/sidebar/Fill/CloseIconOne';
+import SearchIcon from '../../images/sidebar/Outline/Search';
+import SettingsIcon from '../../images/sidebar/Outline/Settings';
+import PersonIcon from '../../images/sidebar/Outline/Person';
+import PowerIcon from '../../images/sidebar/Outline/Power';
 import s from './Header.module.scss';
 
 class Header extends React.Component {
@@ -105,27 +112,44 @@ class Header extends React.Component {
 
     return (
       <Navbar className={`${s.root} d-print-none ${navbarType === NavbarTypes.FLOATING ? s.navbarFloatingType : ''}`} style={{backgroundColor: navbarColor, zIndex: !openUsersList ? 100 : 0}}>
-        <Nav>
+       <Nav>
           <NavItem>
-            <NavLink className="d-md-down-none ml-5" id="toggleSidebar" onClick={this.toggleSidebar}>
-            <i className={`la la-bars ${chroma(navbarColor).luminance() < 0.4 ? "text-white" : ""}`}/>
+            <NavLink className={`d-md-down-none ml-5 ${s.toggleSidebar}`} id="toggleSidebar" onClick={this.toggleSidebar}>
+              <span className={s.headerSvgFlipColor}>
+                <MenuIcon  maskId={1001}/>
+              </span>
             </NavLink>
             <UncontrolledTooltip placement="bottom" target="toggleSidebar">
               Turn on/off<br />sidebar<br />collapsing
             </UncontrolledTooltip>
             <NavLink className="fs-lg d-lg-none" onClick={this.switchSidebar}>
-            <span
-              className={`rounded rounded-lg d-md-none d-sm-down-block`}>
-                <i 
-                  className="la la-bars" 
-                  style={{fontSize: 30, color: navbarColor === "#ffffff" 
-                  ? "#ffffff"
-                  : chroma(navbarColor).luminance() < 0.4 ? "#ffffff" : ""}} 
-                />
+            <span 
+              className={`rounded rounded-lg d-md-none d-sm-down-block`} style={{marginTop: 7}}>
+                <span 
+                  className={s.headerSvgFlipColor}
+                  style={{fontSize: 30}} 
+                ><MenuIcon maskId={1000}/></span>
               </span>
-              <i className={`la la-bars ml-3 d-sm-down-none ${chroma(navbarColor).luminance() < 0.4 ? "text-white" : ""}`}/>
+              <span className={`ml-3 d-sm-down-none ${s.headerSvgFlipColor}`}>
+                <MenuIcon maskId={999}/>
+              </span>
             </NavLink>
           </NavItem>
+          <NavItem className="d-sm-down-none">
+            <NavLink className="px-2">
+              <span className={s.headerSvgFlipColor}>
+                <FlipIcon />
+              </span>
+            </NavLink>
+          </NavItem>
+          <NavItem className="d-sm-down-none">
+            <NavLink className="px-2">
+              <span className={s.headerSvgFlipColor}>
+                <CloseIcon />
+              </span>
+            </NavLink>
+          </NavItem>
+
         </Nav>
 
         <Form className={`d-sm-down-none ml-5 ${s.headerSearchInput}`} inline>
@@ -134,14 +158,14 @@ class Header extends React.Component {
               cx('input-group-no-border', {'focus' : !!focus})
             }>
               <InputGroupAddon addonType="prepend">
-                <i className="la la-search" />
+              <span className={`${s.headerSvgFlipColor}`}><SearchIcon /></span>
               </InputGroupAddon>
               <Input id="search-input" placeholder="Search Dashboard" className={cx({'focus' : !!focus})} />
             </InputGroup>
           </FormGroup>
         </Form>
 
-        <NavLink className={`${s.navbarBrand} d-md-none ${chroma(navbarColor).luminance() < 0.4 ? "text-white" : ""}`}>
+        <NavLink  className={`${s.navbarBrand} d-md-none ${s.headerSvgFlipColor}`}>
           <i className="la la-circle text-primary mr-n-sm" />
           <i className="la la-circle text-danger" />
           &nbsp;
@@ -164,12 +188,14 @@ class Header extends React.Component {
             <span className={`d-sm-down-none ${chroma(navbarColor).luminance() < 0.4 ? "text-white" : ""}`}>{user && (user.firstName || user.email)}</span>
           </NavbarText>
           <Dropdown nav isOpen={this.state.menuOpen} toggle={this.toggleMenu} className="tutorial-dropdown pr-4">
-            <DropdownToggle nav>
-            <i className={`la la-cog ${chroma(navbarColor).luminance() < 0.4 ? "text-white" : ""}`} />
+            <DropdownToggle nav className={`${s.mobileCog}`}>
+              <span className={`${s.headerSvgFlipColor}`}>
+                <SettingsIcon/>
+              </span>
             </DropdownToggle>
-            <DropdownMenu right className={`super-colors`}>
-              <DropdownItem href="/#/app/profile"><i className="la la-user" /> My Account</DropdownItem>
-              <DropdownItem onClick={this.doLogout}><i className="la la-sign-out" /> Log Out</DropdownItem>
+            <DropdownMenu right className={`${s.headerDropdownLinks} super-colors`}>
+              <DropdownItem href="/#/app/profile"><span className={s.headerDropdownIcon}><PersonIcon/></span> My Account</DropdownItem>
+              <DropdownItem onClick={this.doLogout}><span className={s.headerDropdownIcon}><PowerIcon/></span> Log Out</DropdownItem>
             </DropdownMenu>
           </Dropdown>
         </Nav>